Documentation, argparse, typo fixes
