+++ download/manager.py
@@ -7,9 +7,9 @@ from download.sources.spigotmc_source import SpigotMcSource
 from download.sources.jenkins_source import JenkinsSource
 from download.sources.github_source import GitHubSource
 
-from download.postprocessors.paperclip_postprocessor import PaperClipPostprocessor
-from download.postprocessors.plugin_postprocessor import PluginPostprocessor
-from download.postprocessors.zip_postprocessor import ZipPostprocessor
+from download.post_processors.paperclip_postprocessor import PaperClipPostprocessor
+from download.post_processors.plugin_postprocessor import PluginPostprocessor
+from download.post_processors.zip_postprocessor import ZipPostprocessor
 
 from download.destinations.basic_destination import BasicDestination
 from download.destinations.s3_destination import S3Destination
@@ -39,19 +39,22 @@ class DownloadManager():
         self.basic_destination = BasicDestination()
         self.s3_destination = S3Destination()
 
-    def download(self, source, name, url, postprocessors, **kwargs):
+    def download(self, source, name, url, post_processors, **kwargs):
+        # Download file from the right source
         source = self.get_source_manager(source)
         downloaded_binary = source.download_element(url, **kwargs)
 
-        for postprocessor in postprocessors:
+        # Run post_processors
+        for postprocessor in post_processors:
             processor = self.get_postprocessing_manager(postprocessor)
             downloaded_binary, source, name, url = processor.process(downloaded_binary, source, name, url, **kwargs)
 
+        # Save plugin somewhere
         destination = self.get_destination_manager()
         destination.save(downloaded_binary, source, name, url, **kwargs)
 
     def download_resources(self, resources):
-        for resource in tqdm(resources):
+        for resource in tqdm(resources, desc='Downloading resources'):
             try:
                 self.download(**resource)
             except Exception:
@@ -69,6 +72,7 @@ class DownloadManager():
         return self.__getattr_safe('destination', destination)
 
     def __getattr_safe(self, resource_name, value):
+        # A simple wrapper to get our resources dynamically
         try:
             return getattr(self, '{}_{}'.format(value, resource_name))
         except AttributeError:
