+++ download/sources/jenkins_source.py
@@ -17,11 +17,13 @@ class JenkinsSource(DirectSource):
 
         stripped_url = url.strip('/')
 
+        # Get the last successful build
         jenkins_json_url = '{}/lastSuccessfulBuild/api/json'.format(stripped_url)
 
         jenkins_response = self.session.get(jenkins_json_url).json()
 
         for artifact in jenkins_response.get('artifacts') or []:
+            # Find and return the file from the build
             if filter_regex.match(artifact['fileName']):
                 artifact_url = '{}/lastSuccessfulBuild/artifact/{}'.format(stripped_url, artifact['relativePath'])
