+++ redcraft_plugin_updater.py
@@ -1,4 +1,5 @@
 import json
+import argparse
 from dotenv import load_dotenv
 
 from download.manager import DownloadManager
@@ -7,9 +8,17 @@ from download.manager import DownloadManager
 if __name__ == '__main__':
     load_dotenv()
 
-    with open('resources_to_download.json') as resources_file:
-        resources = json.load(resources_file)
+    parser = argparse.ArgumentParser(description='Process some integers.')
+    parser.add_argument(
+        'plugin_json_file',
+        type=argparse.FileType('r'),
+        help='The JSON file with the list of the plugins to download'
+    )
 
-        download_manager = DownloadManager()
+    args = parser.parse_args()
 
-        download_manager.download_resources(resources)
+    # Parse resources
+    resources = json.load(args.plugin_json_file)
+
+    # Download resources
+    DownloadManager().download_resources(resources)
