+++ redcraft_plugins.json
@@ -3,187 +3,187 @@
         "name": "Yatopia",
         "url": "https://ci.codemc.io/job/YatopiaMC/job/Yatopia/job/ver%252F1.16.4/",
         "source": "jenkins",
-        "postprocessors": ["paperclip"],
+        "post_processors": ["paperclip"],
         "filter": "yatopia-*.jar"
     },
     {
         "name": "CMI",
         "url": "https://www.spigotmc.org/resources/cmi-270-commands-insane-kits-portals-essentials-economy-mysql-sqlite-much-more.3742/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "Panilla",
         "url": "https://www.spigotmc.org/resources/panilla-prevent-hacked-items.65694/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "Head Database",
         "url": "https://www.spigotmc.org/resources/head-database.14280/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "AAC",
         "url": "https://www.spigotmc.org/resources/aac-advanced-anti-cheat-hack-kill-aura-blocker.6442/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "Reconnect",
         "url": "https://www.spigotmc.org/resources/bungee-reconnect.16429/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "LiteBans",
         "url": "https://www.spigotmc.org/resources/litebans.3715/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "goPaint",
         "url": "https://www.spigotmc.org/resources/gopaint.27717/",
         "source": "spigotmc",
-        "postprocessors": ["zip", "plugin"],
+        "post_processors": ["zip", "plugin"],
         "archive_filter": "goPaint 1.14*.jar"
     },
     {
         "name": "goBrush",
         "url": "https://www.spigotmc.org/resources/gobrush.23118/",
         "source": "spigotmc",
-        "postprocessors": ["zip", "plugin"],
+        "post_processors": ["zip", "plugin"],
         "archive_filter": "goBrush 1.14*.jar"
     },
     {
         "name": "FarmLimiter",
         "url": "https://www.spigotmc.org/resources/farm-limiter.1419/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "SelectionVisualizer",
         "url": "https://www.spigotmc.org/resources/selection-visualizer.22631/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "PlotSquared",
         "url": "https://www.spigotmc.org/resources/plotsquared-v5.77506/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "EpicWorldGenerator",
         "url": "https://www.spigotmc.org/resources/epicworldgenerator-1-15-1-16-4-%E2%9D%97%EF%B8%8Fsale-%E2%9D%97%EF%B8%8F-20-off.8067/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "ClientStats",
         "url": "https://www.spigotmc.org/resources/clientstats-bungee.27919/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "RedstoneClockDetector",
         "url": "https://www.spigotmc.org/resources/redstoneclockdetector.18475/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "BungeePluginManager",
         "url": "https://www.spigotmc.org/resources/bungeepluginmanager.7288/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "WorldBorder",
         "url": "https://www.spigotmc.org/resources/worldborder-1-15.80466/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "F3NPerm",
         "url": "https://www.spigotmc.org/resources/f3nperm.46461/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "BookExploitFix",
         "url": "https://www.spigotmc.org/resources/bookexploitfix.5897/",
         "source": "spigotmc",
-        "postprocessors": ["plugin"]
+        "post_processors": ["plugin"]
     },
     {
         "name": "ServerListPlus",
         "url": "https://ci.codemc.io/job/Minecrell/job/ServerListPlus/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "ServerListPlus-*-Universal.jar"
     },
     {
         "name": "FastAsyncWorldEdit",
         "url": "https://ci.athion.net/job/FastAsyncWorldEdit-1.16/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "FastAsyncWorldEdit-1.16-*.jar"
     },
     {
         "name": "ViaVersion",
         "url": "https://ci.viaversion.com/job/ViaVersion/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "ViaVersion-*.jar"
     },
     {
         "name": "spark",
         "url": "https://ci.lucko.me/job/spark/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "spark.jar"
     },
     {
         "name": "LuckPerms",
         "url": "https://ci.lucko.me/job/LuckPerms/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "LuckPerms-Bukkit-*.jar"
     },
     {
         "name": "LuckPermsBungee",
         "url": "https://ci.lucko.me/job/LuckPerms/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "LuckPerms-Bungee-*.jar"
     },
     {
         "name": "BungeeTabListPlus",
         "url": "https://ci.codecrafter47.de/job/BungeeTabListPlus/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "BungeeTabListPlus-*.jar"
     },
     {
         "name": "BungeeTabListPlusBridge",
         "url": "https://ci.codecrafter47.de/job/BungeeTabListPlus/",
         "source": "jenkins",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "BungeeTabListPlus_BukkitBridge-*.jar"
     },
     {
         "name": "RedCraftReboot",
         "url": "https://github.com/redcraft-org/RedCraftReboot",
         "source": "github",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "RedCraftReboot-*.jar"
     },
     {
         "name": "RedCraftBungeeJsonApi",
         "url": "https://github.com/redcraft-org/RedCraftBungeeJsonApi",
         "source": "github",
-        "postprocessors": ["plugin"],
+        "post_processors": ["plugin"],
         "filter": "RedCraftBungeeJsonApi-*.jar"
     }
 ]
